+++ spec/index-spec.js
@@ -110,11 +110,25 @@ describe('Builder', () => {
     });
 
     describe('\'--html\'', () => {
-      it('should generate a html file named \'preview.html\' with a preview for all the icons into the generated font', (done) => {
+      it('should generate a html file named \'dcsIconFont-preview.html\' with a preview for all the icons into the generated font', (done) => {
         const options = config.getConfig({ html: true });
         builder.build(options)
           .then((result) => {
-            fs.readFile('./build/preview.html', 'utf8', (err, file) => {
+            fs.readFile('./build/dcsIconFont-preview.html', 'utf8', (err, file) => {
+              expect(file).toBeDefined();
+              expect(result).toBeSuccessful();
+              done();
+            });
+          });
+      });
+    });
+
+    describe('\'--htmlDest\'', () => {
+      it('should generate the html preview with in the path \'build/sample/sample-preview.html\'', (done) => {
+        const options = config.getConfig({ html: true, htmlDest: 'build/sample', fontname: 'sample' });
+        builder.build(options)
+          .then((result) => {
+            fs.readFile('./build/sample/sample-preview.html', 'utf8', (err, file) => {
               expect(file).toBeDefined();
               expect(result).toBeSuccessful();
               done();